cleaned up compute_ng script
